+++ Workspace/Debug/maps/testMap/map.lua
@@ -8,5 +8,9 @@ map = {
 		sizeX = 1024,
 		sizeY = 1024,
 	},
+	lighting = {
+		sun_color = {r=0.2, g=0.2, b=0.2},
+		sun_position = {x=512, y=100, z=512},
+	},
 	texture_gammaCorrection = false,
 }