+++ CHANGELOG.md
@@ -17,6 +17,8 @@
  - Tweaked the output range of maps from the heightmap data to get a more consistent output
  - Added map smoothing to produce a better output
  - Fixed 90 degree rotation of heightmaps
+ - Begun exposing information to the LuaScene
+ - Added sun position and color specificatin in map.lua
 ##### Profiler
  - Changed profiling to output at the end of each frame if applicable, instead of hogging memory in the background
  - Changed frequency from every 20th frame to 200th